+++ app/src/services/coreStateApi.ts
@@ -2,7 +2,7 @@ import type { User } from '../types/api';
 import type { TeamInvite, TeamMember, TeamWithRole } from '../types/team';
 import type { AccessibilityStatus } from '../utils/tauriCommands/accessibility';
 import type { AutocompleteStatus } from '../utils/tauriCommands/autocomplete';
-import type { ServiceStatus } from '../utils/tauriCommands/hardware';
+import type { ServiceStatus } from '../utils/tauriCommands/service';
 import type { LocalAiStatus } from '../utils/tauriCommands/localAi';
 import { callCoreRpc } from './coreRpcClient';
