--- src/openhuman/tools/impl/hardware/mod.rs
@@ -1,7 +1,0 @@
-mod board_info;
-mod memory_map;
-mod memory_read;
-
-pub use board_info::HardwareBoardInfoTool;
-pub use memory_map::HardwareMemoryMapTool;
-pub use memory_read::HardwareMemoryReadTool;
