+++ src/openhuman/tools/impl/mod.rs
@@ -2,7 +2,6 @@ pub mod agent;
 pub mod browser;
 pub mod cron;
 pub mod filesystem;
-pub mod hardware;
 pub mod memory;
 pub mod network;
 pub mod system;
@@ -11,7 +10,6 @@ pub use agent::*;
 pub use browser::*;
 pub use cron::*;
 pub use filesystem::*;
-pub use hardware::*;
 pub use memory::*;
 pub use network::*;
 pub use system::*;
